requirements.txt
pydub==0.25.1
openai==0.28.0
python-dotenv==1.0.0
langchain==0.0.315
datetime==5.4
wikipedia-api==0.5.8
numexpr==2.9.0
pypdf==4.0.2
yt-dlp==2023.12.30
chromadb==0.4.24
SpeechRecognition==3.8.1
openai-whisper==20231117
torch==2.2.1
gTTS==2.5.1
PyAudio==0.2.14
